### 2017-04-07 #233 @whedon
[REVIEW]: nails: Network Analysis Interface for Literature Studies
**Submitting author:** @jsalminen (<a href="http://orcid.org/0000-0003-0688-2211">Juho Salminen</a>)
**Repository:** <a href="https://github.com/aknutas/nails-package" target ="_blank">https://github.com/aknutas/nails-package</a>
**Version:** 1.0.2
