### 2017-04-07 #233 @aknutas
Hello everyone. Test data for the review process is available at https://akpstorage.s3.amazonaws.com/nails_joss_testinput.zip in case any participant to the process does not have access to the web of science.
