### 2017-04-08 #233 @janschulz
> Version: Does the release version given match the GitHub release (1.0.0)?

This is still a rc, so it current does not match -> needs to be addressed...

> Installation and Usage

That was a bit of pain: https://github.com/aknutas/nails/issues/12

Neverthless, it worked, somehow: :white_check_mark:

> Automatic tests

These are missing as far as I can see.. :question:

All in all this is a great "Product" which works great as a webservice (via http://hammer.nailsproject.net/), but the way the repo is structured for users it is a bit unusual and IMO would work much better as a package which could be installed (pulling in all required dependencies) and then used as a function.

> statement of need

IMO a more "user centric" verison would be nice as the current one "undersells" the software a bit: "NAILS performs statistical and Social Network Analysis (SNA) on citation data. SNA is a new way for researchers to map large datasets and get insights from new angles by analyzing connections between articles." is correct, but IMO this is alse a great tool to simply make sense of a topic which is new to me and I want to gain a first insight what is important here and how important topics/authors are connected. Neverthelss: :white_check_mark:

> citations

IMO the packages should be cited (e.g. [igraph](https://cran.r-project.org/web/packages/igraph/citation.html)). See here how to get all citation info: http://www.astrostatistics.psu.edu/su07/R/html/utils/html/citation.html or this here has a way to get all citations of your currently loaded packaes: http://www.carlboettiger.info/2012/03/20/citing-r-packages.html

# All in all

IMO there are two big issues:
*automatic tests* (and the way the code is structured): I'm not sure how to handle automatic tests in the current way the package is structured (rmd instead of functions in a package). IMO it should become a package which has a few API functions which are called by a user and which spit out a html file for a user. These can then be tested. This would also make it a more releaseable package. -> I'm not sure if this is a blocker or not. Someone else should answer that. :-)
* The paper should cite the packages it depends on.
