### 2017-04-10 #233 @karthik
@jsalminen Can you please review and respond to the issues raised above?

thank you.
