### 2017-04-11 #233 @aknutas
Thank you for the feedback, especially with the citation issue. Instructions @janschulz gave us here were very helpful to ensure that the references to other scientific software are up to date.

@karthik, we have addressed the minor and major issues as listed below.

The visual improvement proposals were assigned a responsible person, taken under work and set a future release version (issues aknutas/nails#13, aknutas/nails#14 and aknutas/nails#15).

The major issues were corrected to the best of our ability:

1. The release version was changed to v1.0.0 .
2. We created a start up Linux script for manual testing and GUI-less startup. Changing the software to be an R package would require major refactoring both in the R code and the online web analysis site we provide. However, we discussed the possibility with @jsalminen and are considering using a portion of our upcoming summer vacations to create the package. (issue aknutas/nails#12)
3. Citations to all packages we use were added.
