### 2017-04-13 #233 @karthik
@jsalminen thank you so much for the thoughtful review, we really appreciate it. And thanks also to @aknutas for the submission and timely responses.

I apologize for not noticing earlier (prior to inviting review) that the submission isn't a software package, but a collection of scripts. As a journal we strongly prefer that software be packaged (and ideally also submitted to a package manager) rather than exist as a collection of notebooks and scripts. This would alleviate issues around automatic testing, and also installation issues (especially with the dependencies). Much of the reviewer checklist is also built around the idea that the submission being reviewed is a package.

However, given that the review is now complete, as a gesture of good will to both the submitting authors and the reviewer we are going to accept this submission but plan to modify our submission guidelines to document that submissions of this kind would not meet our submission criteria in the future.

As @janschulz notes:

>  but the way the repo is structured for users it is a bit unusual and IMO would work much better as a package which could be installed (pulling in all required dependencies) and then used as a function.

We also agree that this would be better off as an R package. I see that the authors note that this refactoring would be a considerable effort on their part, and not something they will be able to do till the summer (issue aknutas/nails#12). So if the authors are willing, we are also happy to pause the review while they invest the effort in creating an actual R package and then follow through with the acceptance immediately after.

@aknutas: Please let us know your thoughts.
