### 2017-04-18 #233 @jsalminen
Thank you @karthik and @janschulz for the review! The suggestion about building an R package is very insightful and we've decided to go for it. As we're not in a rush with the publication, we figured it is better to do it properly, even if it takes a while longer.

You can pause the review and we'll get back to you once the R package is ready.
