### 2017-04-22 afd1f98
Added intitial GPLv3 license files and the readme
