+++ README.md
@@ -6,7 +6,7 @@ at _Lappeenranta University of Technology_
 
 What Is It?
 ----
-This R package provides tools for performing Social Network Analysis (SNA) on citation data. This version is meant to use for R programmers. A general use version to be executed in RStudio is available at the second repository in the alternative [repository repository](https://github.com/aknutas/nails). See more information at the [project page](http://aknutas.github.io/nails/)
+This R package provides tools for performing Social Network Analysis (SNA) on citation data. This version is meant to use for R programmers. A general use version to be executed in RStudio is available at the second repository in the alternative [repository repository](https://github.com/aknutas/nails). See more information at the [project page](http://aknutas.github.io/nails/).
 
 SNA is an interesting way for researchers to map large datasets and get insights from new angles. The steps for downloading data from Web of Knowledge and using our tools to process it are detailed below. The set of tools which are required to perform the analyses are free and need a minimum amount of installation. Furthermore, we will soon make a web-based analysis server HAMMER available so that you can process the data without needing to do any installation or manual processing steps.
 
