### 2017-04-22 29abac5
First draft of the nails package paper for JOSS
Including codemeta file
