+++ man/read_wos_txt.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/read_data.R
+\name{read_wos_txt}
+\alias{read_wos_txt}
+\title{Read a .txt-file downloaded from Web of Science.}
+\usage{
+read_wos_txt(filepath)
+}
+\arguments{
+\item{filepath}{Path to a folder}
+}
+\value{
+A data frame
+}
+\description{
+Read a .txt-file downloaded from Web of Science.
+}
