+++ man/read_wos_folder.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/read_data.R
+\name{read_wos_folder}
+\alias{read_wos_folder}
+\title{Read all .txt-files downloaded from Web of Science within a folder.}
+\usage{
+read_wos_folder(filepath)
+}
+\arguments{
+\item{filepath}{Path to a folder}
+}
+\value{
+A data frame
+}
+\description{
+Read all .txt-files downloaded from Web of Science within a folder.
+}
