### 2017-04-22 af91484
Add clean_data.R and test_clean_data.R.
