### 2017-04-23 138a406
Add cleaning functions, check data by column names.
