+++ R/clean_data.R
@@ -1,4 +1,13 @@
 
+#' Helper function to remove leading and trailing whitespace.
+#' @param x A string
+#' @return A string
+trim <- function (x) gsub("^\\s+|\\s+$", "", x)
+
+#' Helper function to remove quotation marks from strings.
+#' @param df A data frame
+#' @param columns A vector of column names to be processed
+#' @return A data frame
 remove_quotes <- function(df, columns) {
     for (col in columns) {
         df[[col]] <- gsub("'", "", df[[col]])
@@ -7,6 +16,11 @@ remove_quotes <- function(df, columns) {
     return(df)
 }
 
+#' Helper function to change strings to lowercase in
+#' selected data frame columns.
+#' @param df A data frame
+#' @param columns A vector of column names to be processes
+#' @return A data frame
 change_lowercase <- function(df, columns) {
     for (col in columns) {
         df[[col]] <- tolower(df[[col]])
@@ -14,6 +28,11 @@ change_lowercase <- function(df, columns) {
     return(df)
 }
 
+#' Helper function to change strings to uppercase in
+#' selected data frame columns.
+#' @param df A data frame
+#' @param columns A vector of column names to be processes
+#' @return A data frame
 change_uppercase <- function(df, columns) {
     for (col in columns) {
         df[[col]] <- toupper(df[[col]])
@@ -21,7 +40,14 @@ change_uppercase <- function(df, columns) {
     return(df)
 }
 
+#' Reprocess data from WOS for further analysis
+#' @param df A data frame containing data from WOS.
+#' @return A data frame
 clean_wos_data <- function(df) {
+    data_type <- check_data(df)
+    if (data_type == 1) {
+        names(df) <- fixed_fieldtags
+    }
     df <- change_lowercase(df, columns = c("AuthorKeywords",
                                            "KeywordsPlus",
                                            "SubjectCategory"))
@@ -40,9 +66,19 @@ clean_wos_data <- function(df) {
     df$CitedReferences <- gsub("DOI DOI", "DOI", df$CitedReferences)
     df$YearPublished <- as.numeric(df$YearPublished)
     df$TimesCited <- as.numeric(df$TimesCited)
+
+    df$Location <- sapply(df$AuthorAddress, get_location)
+
+    df$ReferenceString <- apply(df, 1, makeRef)
+    df <- df[!duplicated(df[, "ReferenceString"]), ]
+
+    df$ID <- 1:nrow(df)
+    return(df)
 }
 
-# Helper function to construct strings
+#' Helper function to construct reference strings
+#' @param x A row from a data frame containing cleaned data from WOS
+#' @return  A string in same format as references in CitedReferences
 makeRef <- function(x) {
     refstring <- getName(x)
     if (!is.na(x["YearPublished"])) {
@@ -63,7 +99,9 @@ makeRef <- function(x) {
     return(refstring)
 }
 
-# Helper function to extract the name of first author
+#' Helper function to extract the name of first author
+#' @param x A string containing full author names
+#' @return A string
 getName <- function(x) {
     name = NA
     try( {
@@ -80,4 +118,26 @@ getName <- function(x) {
         }
     } )
     return(name)
+}
+
+#' Helper function for extracting countries and cities from AuthorAddress.
+#' @param x A string containing AuthorAddress
+#' @return A string containing city and country
+get_location <- function(x) {
+    country <- NA
+    city <- NA
+    if (x != "") {
+        x <- gsub("\\[.*?\\]", "", x)
+        x <- unlist(strsplit(x, ";"))
+        x <- x[x != " "]
+        cities <- sapply(x, function(x) tail(unlist(strsplit(x, ",")), 2))
[…]
