+++ man/trim.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_data.R
+\name{trim}
+\alias{trim}
+\title{Helper function to remove leading and trailing whitespace.}
+\usage{
+trim(x)
+}
+\arguments{
+\item{x}{A string}
+}
+\value{
+A string
+}
+\description{
+Helper function to remove leading and trailing whitespace.
+}
