+++ man/change_lowercase.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_data.R
+\name{change_lowercase}
+\alias{change_lowercase}
+\title{Helper function to change strings to lowercase in
+selected data frame columns.}
+\usage{
+change_lowercase(df, columns)
+}
+\arguments{
+\item{df}{A data frame}
+
+\item{columns}{A vector of column names to be processes}
+}
+\value{
+A data frame
+}
+\description{
+Helper function to change strings to lowercase in
+selected data frame columns.
+}
