+++ man/clean_wos_data.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_data.R
+\name{clean_wos_data}
+\alias{clean_wos_data}
+\title{Reprocess data from WOS for further analysis}
+\usage{
+clean_wos_data(df)
+}
+\arguments{
+\item{df}{A data frame containing data from WOS.}
+}
+\value{
+A data frame
+}
+\description{
+Reprocess data from WOS for further analysis
+}
