+++ man/get_location.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_data.R
+\name{get_location}
+\alias{get_location}
+\title{Helper function for extracting countries and cities from AuthorAddress.}
+\usage{
+get_location(x)
+}
+\arguments{
+\item{x}{A string containing AuthorAddress}
+}
+\value{
+A string containing city and country
+}
+\description{
+Helper function for extracting countries and cities from AuthorAddress.
+}

[…]
