### 2017-06-11 1be81ae
Add functions to create a citation network.
