+++ R/clean_data.R
@@ -69,18 +69,18 @@ clean_wos_data <- function(df) {
 
     df$Location <- sapply(df$AuthorAddress, get_location)
 
-    df$ReferenceString <- apply(df, 1, makeRef)
+    df$ReferenceString <- apply(df, 1, make_reference)
     df <- df[!duplicated(df[, "ReferenceString"]), ]
 
-    df$ID <- 1:nrow(df)
+    df$id <- 1:nrow(df)
     return(df)
 }
 
 #' Helper function to construct reference strings
 #' @param x A row from a data frame containing cleaned data from WOS
 #' @return  A string in same format as references in CitedReferences
-makeRef <- function(x) {
-    refstring <- getName(x)
+make_reference <- function(x) {
+    refstring <- get_name(x)
     if (!is.na(x["YearPublished"])) {
         refstring <- paste(refstring, x["YearPublished"], sep = ", ")
     }
@@ -102,7 +102,7 @@ makeRef <- function(x) {
 #' Helper function to extract the name of first author
 #' @param x A string containing full author names
 #' @return A string
-getName <- function(x) {
+get_name <- function(x) {
     name = NA
     try( {
         names <- unlist(strsplit(x["AuthorFullName"], ";"))
@@ -120,6 +120,28 @@ getName <- function(x) {
     return(name)
 }
 
+#' Helper function to extract DOIs
+#' @param x A list of strings containing DOI in second element
+#' @return A string cointaining DOI
+get_DOI <- function(x) {
+    DOI <- NA
+    if (length(x) == 2) {
+        DOI <- x[2]
+    }
+    return(DOI)
+}
+
+#' Helper function to extract years
+#' @param x A list of strings containing year in the second element
+#' @return A year number
+get_year <- function(x) {
+    year <- NA
+    if (length(x) > 1) {
+        year <- as.numeric(x[2])
+    }
+    return(year)
+}
+
 #' Helper function for extracting countries and cities from AuthorAddress.
 #' @param x A string containing AuthorAddress
 #' @return A string containing city and country
@@ -135,9 +157,177 @@ get_location <- function(x) {
         city <- sapply(city, trim)
         #   country <- gsub(" ", "", cities[2, ])
         country <- sapply(cities[2, ], trim)
-        return(paste(paste(city, country, sep = ", "), collapse = ";"))
+        location <- paste(paste(city, country, sep = ", "), collapse = ";")
     }
     else {
-        return(NA)
+        location <- NA
     }
+    return(location)
 }
+
+
+#' Organise literature data by selected column
+#' @param df A data frame containing the literature data
+#' @param selection A string with a column name
+#' @param split A character used to split selection
+#' @return A data frame with literature organised by selected column
+arrange_by <- function(df, selection, sep = ";") {
+    # Check data
+
+    # Subset data
+    # POSSIBLE OPTIMIZATION: remove this step!
+    arranged_df <- subset(df, select = c(selection, "id"))
+    # Remove NAs
+    arranged_df <- arranged_df[!is.na(arranged_df[, selection]), ]
+    # Create data frame: selection split by split, each element on a new row,
+    # id copied to new rows
+    arranged_df <- splitstackshape::cSplit(arranged_df,
+                                         splitCols = selection,
+                                         sep = sep, direction = "long")
+    temp <- quote(selection)
+    arranged_df <- arranged_df[,  eval(temp):= as.character(get(selection))]
+    # Removing rows with NA as author name created in previous step
+    arranged_df <- arranged_df[!is.na(arranged_df[, selection]), ]
+    # Merge the rest of the data by id
+    arranged_df <- merge(arranged_df,
+                       df[, !(names(df) %in% c(selection))],
[…]
