### 2017-06-11 c98e85b
Added Travis CI support and travis shield
