### 2017-06-11 4278b69
Added baseline topicmodeling code
Tests not implemented yet. (TODO)

Will be merged to main codebase when verified to work.
