+++ tests/testthat/test_topic_model.R
@@ -1,0 +1,25 @@
+library(nails)
+context("Building topic models")
+
+load("data/fieldtags.RData")
+load("data/column_names.RData")
+
+test_that("Topicmodels input data gets cleaned properly", {
+    # Not implemented yet
+    expect_true(FALSE)
+})
+
+test_that("stm corpus gets built", {
+    # Not implemented yet
+    expect_true(FALSE)
+})
+
+test_that("topicmodels corpus gets built", {
+    # Not implemented yet
+    expect_true(FALSE)
+})
+
+test_that("topicmodels dtm gets built", {
+    # Not implemented yet
+    expect_true(FALSE)
+})
