+++ R/topic_model.R
@@ -1,0 +1,193 @@
+#' Estimates number of topics (K) using the stm library
+#'
+#' @param literature The cleaned and processed literature dataframe
+#' @return A list with vector of words (textlist) and data frame with only abstracts, topics and rowids (doctablewt)
+preprocess_literature_for_topicmodeling <- function(literature) {
+    if(missing(literature))
+        stop("Need valid literature input")
+
+    # Insert rowids for output merging
+    literature$topicmodelrowids <-as.numeric(rownames(literature))
+
+    # Subselect dataframe for processing and drop empty rows
+    doctablewt <- literature[,c("DocumentTitle", "Abstract", "topicmodelrowids")]
+    doctablewt <- doctablewt[!is.na(doctablewt$Abstract), ]
+    doctablewt <- doctablewt[!is.na(doctablewt$DocumentTitle), ]
+
+    # Extract and combine topics, abstracts
+    data <- paste(doctablewt$DocumentTitle, doctablewt$Abstract, sep = " ")
+    return(list(textlist = data, doctablewt = doctablewt))
+}
+
+#' Create stm library corpus
+#'
+#' @param textlist Vector of processed words
+#' @return Corpus for stm
+build_stm_corpus <- function(textlist) {
+    # Prepare documents into corpus
+    # Also processes the following by default:
+    # Lowercases, removes SMART stopwords, removes numbers, removes punctuation, wordlength min is 3
+    processed <- textProcessor(textlist, stem = TRUE)
+    out <- prepDocuments(processed$documents, processed$vocab)
+    return(out)
+}
+
+#' Estimates number of topics (K) using the stm library
+#'
+#' @param data Data from preprocess_literature_for_topicmodeling()
+#' @return A list with topic estimation from stm lib (topicest), semantic coherence values (semcohvalues) and estimate of best K of topics (bestk)
+select_optimal_k <- function(data) {
+    out <- build_stm_corpus(data$textlist)
+
+    # Estimate number of topics; semantic coherence often good (default method spectral; best compromise and deterministic)
+    # Seed set for consitent results
+    set.seed(5707363)
+    topickest <- searchK(out$documents, out$vocab, K = c(4:12), seed = 5707363)
+    semcohsK <- data.frame(topickest$results$K, topickest$results$semcoh)
+    colnames(semcohsK)<- c("K","semcohs")
+
+    # Getting the K with highest semantic coherence, setting it as true
+    bestpick <- semcohsK[which.max(semcohsK$semcohs),]
+    bestK <- as.integer(bestpick$K)
+    semcohsK$bestpick <- FALSE
+    semcohsK$bestpick[which.max(semcohsK$semcohs)] <- TRUE
+
+    return(list(topicest = topicest, semcohvalues = semcohsK, bestk = bestpick))
+}
+
+# Function for converting topicmodels to LDAvis compatible
+# Source: https://gist.github.com/a-paxton/a1609f5f772b642027d4
+#' Convert the output of a topicmodels Latent Dirichlet Allocation to JSON
+#' for use with LDAvis
+#'
+#' @param fitted Output from a topicmodels \code{LDA} model.
+#' @param corpus Corpus object used to create the document term
+#' matrix for the \code{LDA} model. This should have been create with
+#' the tm package's \code{Corpus} function.
+#' @param doc_term The document term matrix used in the \code{LDA}
+#' model. This should have been created with the tm package's
+#' \code{DocumentTermMatrix} function.
+#'
+#' @seealso \link{LDAvis}.
+#' @export
+
+topicmodels_json_ldavis <- function(fitted, corpus, doc_term){
+
+    # Find required quantities
[…]
