+++ man/preprocess_literature_for_topicmodeling.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/topic_model.R
+\name{preprocess_literature_for_topicmodeling}
+\alias{preprocess_literature_for_topicmodeling}
+\title{Estimates number of topics (K) using the stm library}
+\usage{
+preprocess_literature_for_topicmodeling(literature)
+}
+\arguments{
+\item{literature}{The cleaned and processed literature dataframe}
+}
+\value{
+A list with vector of words (textlist) and data frame with only abstracts, topics and rowids (doctablewt)
+}
+\description{
+Estimates number of topics (K) using the stm library
+}
