+++ man/make_reference.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/clean_data.R
-\name{makeRef}
-\alias{makeRef}
+\name{make_reference}
+\alias{make_reference}
 \title{Helper function to construct reference strings}
 \usage{
-makeRef(x)
+make_reference(x)
 }
 \arguments{
 \item{x}{A row from a data frame containing cleaned data from WOS}
