+++ man/get_literature_nodes.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_data.R
+\name{get_literature_nodes}
+\alias{get_literature_nodes}
+\title{Helper function for extracting citation network nodes from literature data}
+\usage{
+get_literature_nodes(df)
+}
+\arguments{
+\item{df}{A data frame containing the literature data}
+}
+\value{
+A data frame containing citation nodes from literature data
+}
+\description{
+Helper function for extracting citation network nodes from literature data
+}
