+++ man/get_name.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/clean_data.R
-\name{getName}
-\alias{getName}
+\name{get_name}
+\alias{get_name}
 \title{Helper function to extract the name of first author}
 \usage{
-getName(x)
+get_name(x)
 }
 \arguments{
 \item{x}{A string containing full author names}
