+++ man/get_reference_nodes.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_data.R
+\name{get_reference_nodes}
+\alias{get_reference_nodes}
+\title{Helper function for extracting citation network nodes from reference list}
+\usage{
+get_reference_nodes(reference_list)
+}
+\arguments{
+\item{reference_list}{A data frame containing reference list}
+}
+\value{
+A data frame containing citation nodes from reference
+}
+\description{
+Helper function for extracting citation network nodes from reference list
+}
