### 2017-06-11 0a855da
Fixed typo in DESCRIPTION
GPLv3 -> GPL-3
