+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 0.1.0
 Author: Juho Salminen, Antti Knutas and Arash Hajikhani
 Maintainer: To Be Decided <yourself@somewhere.net>
 Description: This R package provides tools for performing Social Network Analysis (SNA) on citation data. This version is meant to use for R programmers. A general use version to be executed in RStudio is available at the second repository in the alternative repository https://github.com/aknutas/nails. See more information at http://aknutas.github.io/nails/.
-License: GPLv3
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Suggests: testthat
