### 2017-06-12 abfe201
Topicmodeling topic number estimation (K) works
