+++ R/topic_model.R
@@ -1,7 +1,11 @@
-#' Estimates number of topics (K) using the stm library
+#' Preprocess literature file for topicmodeling
+#' Use for select_optimal_k or others that take the "data$textlist" object.
 #'
 #' @param literature The cleaned and processed literature dataframe
 #' @return A list with vector of words (textlist) and data frame with only abstracts, topics and rowids (doctablewt)
+#'
+#' @export
+
 preprocess_literature_for_topicmodeling <- function(literature) {
     if(missing(literature))
         stop("Need valid literature input")
@@ -23,26 +27,30 @@ preprocess_literature_for_topicmodeling <- function(literature) {
 #'
 #' @param textlist Vector of processed words
 #' @return Corpus for stm
+
 build_stm_corpus <- function(textlist) {
     # Prepare documents into corpus
     # Also processes the following by default:
     # Lowercases, removes SMART stopwords, removes numbers, removes punctuation, wordlength min is 3
-    processed <- textProcessor(textlist, stem = TRUE)
-    out <- prepDocuments(processed$documents, processed$vocab)
+    processed <- stm::textProcessor(textlist, stem = TRUE)
+    out <- stm::prepDocuments(processed$documents, processed$vocab)
     return(out)
 }
 
 #' Estimates number of topics (K) using the stm library
 #'
 #' @param data Data from preprocess_literature_for_topicmodeling()
-#' @return A list with topic estimation from stm lib (topicest), semantic coherence values (semcohvalues) and estimate of best K of topics (bestk)
+#' @return A list with topic estimation from stm lib (topickest), semantic coherence values (semcohvalues) and estimate of best K of topics (bestk)
+#'
+#' @export
+
 select_optimal_k <- function(data) {
     out <- build_stm_corpus(data$textlist)
 
     # Estimate number of topics; semantic coherence often good (default method spectral; best compromise and deterministic)
     # Seed set for consitent results
     set.seed(5707363)
-    topickest <- searchK(out$documents, out$vocab, K = c(4:12), seed = 5707363)
+    topickest <- stm::searchK(out$documents, out$vocab, K = c(4:12), seed = 5707363)
     semcohsK <- data.frame(topickest$results$K, topickest$results$semcoh)
     colnames(semcohsK)<- c("K","semcohs")
 
@@ -52,7 +60,7 @@ select_optimal_k <- function(data) {
     semcohsK$bestpick <- FALSE
     semcohsK$bestpick[which.max(semcohsK$semcohs)] <- TRUE
 
-    return(list(topicest = topicest, semcohvalues = semcohsK, bestk = bestpick))
+    return(list(topickest = topickest, semcohvalues = semcohsK, bestk = bestpick))
 }
 
 # Function for converting topicmodels to LDAvis compatible
@@ -100,17 +108,18 @@ topicmodels_json_ldavis <- function(fitted, corpus, doc_term){
 #'
 #' @param textlist Vector of texts
 #' @return Corpus for topicmodels
+
 build_topicmodels_corpus <- function(textlist) {
     # Create corpus
-    abstractCorpus <- Corpus(VectorSource(textlist))
+    abstractCorpus <- tm::Corpus(tm::VectorSource(textlist))
 
     # Preprocess by lowercasing, removing punctuation, numbers, whitespace and stopwords, and finally stemming
-    abstractCorpus <- tm_map(abstractCorpus, content_transformer(tolower))
-    abstractCorpus <- tm_map(abstractCorpus, removePunctuation)
-    abstractCorpus <- tm_map(abstractCorpus, removeNumbers)
[…]
