+++ NAMESPACE
@@ -1,3 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(build_topicmodel_from_literature)
+export(preprocess_literature_for_topicmodeling)
+export(select_optimal_k)
 export(topicmodels_json_ldavis)
