### 2017-06-12 334f6a7
Working topicmodels base files and unit tests with test_That
