+++ man/build_topicmodels_DTM.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/topic_model.R
-\name{build_topicmodels_DTM}
-\alias{build_topicmodels_DTM}
-\title{Create topicmodels library DocumentTermMatrix}
-\usage{
-build_topicmodels_DTM(abstractCorpus)
-}
-\arguments{
-\item{abstractCorpus}{From build_topicmodels_corpus()}
-}
-\value{
-Corpus for topicmodels
-}
-\description{
-Create topicmodels library DocumentTermMatrix
-}
