### 2017-06-12 8eee2fe
Added configuration fixed for Travis-CI and codecov features
