+++ DESCRIPTION
@@ -16,5 +16,6 @@ Imports:
     topicmodels,
     dplyr,
     stringi
-Suggests: testthat
+Suggests:
+    testthat
 RoxygenNote: 6.0.1
