+++ .travis.yml
@@ -3,3 +3,14 @@
 language: R
 sudo: false
 cache: packages
+
+addons:
+  apt:
+    packages:
+    - gsl-bin
+
+r_packages:
+  - covr
+
+after_success:
+  - Rscript -e 'library(covr); codecov()'
