### 2017-06-12 926c935
Added more required packages for R library compilation
