### 2017-06-12 3e4f73f
Updated documentation and testing data
