+++ DESCRIPTION
@@ -3,9 +3,9 @@ Type: Package
 Title: Network Analysis Interface for Literature Studies
 Version: 0.1.0
 Author: Juho Salminen, Antti Knutas and Arash Hajikhani
-Maintainer: To Be Decided <yourself@somewhere.net>
+Maintainer: To Be Decided <maintainer@example.com>
 Description: This R package provides tools for performing Social Network Analysis (SNA) on citation data. This version is meant to use for R programmers. A general use version to be executed in RStudio is available at the second repository in the alternative repository https://github.com/aknutas/nails. See more information at http://aknutas.github.io/nails/.
-License: GPL-3
+License: GPL-3 | file LICENSE
 Encoding: UTF-8
 LazyData: true
 Imports:
@@ -15,7 +15,9 @@ Imports:
     igraph,
     topicmodels,
     dplyr,
-    stringi
+    stringi,
+    plyr,
+    splitstackshape
 Suggests:
     testthat
 RoxygenNote: 6.0.1
