### 2017-06-12 4ace769
Create README.md
Updated package requirements info.
