### 2017-06-14 75b3ec9
Fixing CI tests (#2)
Travis CI runs now without stopping errors. Merging back to main. Changes:
* Avoiding directory name conflicts
* Testing an approach to fix test case .RData loading
* Added topicmodeltest .RData file
* Adjusted Travis settings: warnings_are_errors: true -> false
