### 2017-06-14 9fd5f32
Added code coverage badge to README.md
