### 2017-07-01 0b2de49
Add functions to create author network.
