### 2017-07-02 7297bc1
Add update_fieldtags.R for automatically updating sysdata.Rda. Breaks check_data().
