### 2017-07-02 eaf0617
Fix regression bug in check_data.
