### 2017-07-11 88f3281
Add comments to update_fieldtags.R.
