### 2017-07-11 14eb682
Add option to use unchanged colnames, update fixed colnames.
