### 2017-07-11 79ea3ea
Refactor: move author and citation network functions to their own files.
