+++ R/create_citation_network.R
@@ -1,0 +1,124 @@
+
+
+#' Helper function for extracting reference lists
+#' @param df A data frame containing literature data
+#' @return A data frame containing each reference in separate row
+get_reference_list <- function(df) {
+    rl <- arrange_by(df, "CitedReferences")
+    names(rl)[names(rl) == 'CitedReferences'] <- "FullReference"
+
+    year <- strsplit(rl$FullReference, ",")
+    rl$ReferenceYear <- sapply(year, get_year)
+    references <- strsplit(rl$FullReference, " DOI ")
+    rl$Reference <- sapply(references, get_DOI)
+    rl$Reference[is.na(rl$Reference)] <- rl$FullReference[is.na(rl$Reference)]
+    return(rl)
+}
+
+#' Helper function for extracting citation network nodes from reference list
+#' @param reference_list A data frame containing reference list
+#' @return A data frame containing citation nodes from reference
+get_reference_nodes <- function(reference_list) {
+    citation_nodes <- data.frame(Id = reference_list$Reference,
+                                 YearPublished = reference_list$ReferenceYear,
+                                 FullReference = reference_list$FullReference,
+                                 id = NA,
+                                 PublicationType = NA,
+                                 AuthorFullName = NA,
+                                 DocumentTitle = NA,
+                                 PublicationName = NA,
+                                 BookSeriesTitle = NA,
+                                 Language = NA,
+                                 DocumentType = NA,
+                                 ConferenceTitle = NA,
+                                 ConferenceDate = NA,
+                                 ConferenceLocation = NA,
+                                 ConferenceSponsors = NA,
+                                 AuthorKeywords = NA,
+                                 SubjectCategory = NA,
+                                 TimesCited = NA,
+                                 Abstract = NA,
+                                 DOI = NA,
+                                 Origin = "reference",
+                                 stringsAsFactors = FALSE)
+
+    index <- duplicated(citation_nodes$Id)
+    citation_nodes <- citation_nodes[!index, ]
+    return(citation_nodes)
+}
+
+#' Helper function for extracting citation network nodes from literature data
+#' @param df A data frame containing the literature data
+#' @return A data frame containing citation nodes from literature data
+get_literature_nodes <- function(df) {
+    literature_nodes <- subset(df,
+                               select = c(DOI,
+                                          YearPublished,
+                                          ReferenceString,
+                                          id,
+                                          PublicationType,
+                                          AuthorFullName,
+                                          DocumentTitle,
+                                          PublicationName,
+                                          BookSeriesTitle,
+                                          Language,
+                                          DocumentType,
+                                          ConferenceTitle,
+                                          ConferenceDate,
+                                          ConferenceLocation,
+                                          ConferenceSponsors,
+                                          AuthorKeywords,
+                                          SubjectCategory,
+                                          TimesCited,
+                                          Abstract,
+                                          DOI))
+
+    names(literature_nodes)[c(1, 3, 20)] <- c("Id",
+                                              "FullReference",
+                                              "DOI")
+    index <- literature_nodes$Id == ""
+    literature_nodes$Id[index] <- literature_nodes$FullReference[index]
+    literature_nodes$Origin <- "literature"
+
+    index <- duplicated(literature_nodes)
+    literature_nodes <- literature_nodes[!index, ]
+    return(literature_nodes)
+}
+
+#' Extract nodes for a citation network from literature data
+#' @param df A data frame containing literature data
+#' @param reference_list A data frame containing reference list
+#' @return A data frame containing nodes for a citation network
[…]
