### 2017-07-11 d461f23
Add tests for create_citation_network, remove broken tests from test_read_data.R.
