+++ tests/testthat/test_create_citation_network.R
@@ -1,0 +1,148 @@
+library(nails)
+context("Creating a citation network")
+
+# Read and clean test data
+df <- read_wos_txt(file.path("test_data/savedrecs-mactab.txt"), fix_names = TRUE)
+df <- clean_wos_data(df)
+
+# Pick 3 rows for further processing
+df <- df[c(3, 36, 100), ]
+
+# Reference data
+number_of_rows <- sum(df$CitedReferenceCount)
+reference_1 <- "BARRINGTON L, 2011, ANN GEOPHYS-ITALY, V54, P680, DOI 10.4401/AG-5324"
+reference_43 <- "WANG C., 2011, P 17 ACM SIGKDD INT, P448, DOI 10.1145/2020408.2020480"
+reference_44 <- "ALDUNCE P, 2014, DISASTER PREV MANAG, V23, P252, DOI 10.1108/DPM-07-2013-0130"
+reference_148 <- "ZHANG T., 2009, P 17 ACM INT C MULT, P597, DOI 10.1145/1631272.1631365"
+
+year_1 <- 2011
+year_43 <- 2011
+year_44 <- 2014
+year_148 <- 2009
+
+DOI_1 <- "10.4401/AG-5324"
+DOI_43 <- "10.1145/2020408.2020480"
+DOI_44 <- "10.1108/DPM-07-2013-0130"
+DOI_148 <- "10.1145/1631272.1631365"
+
+nodes_colnames <- c("Id",
+                    "YearPublished",
+                    "FullReference",
+                    "id",
+                    "PublicationType",
+                    "AuthorFullName",
+                    "DocumentTitle",
+                    "PublicationName",
+                    "BookSeriesTitle",
+                    "Language",
+                    "DocumentType",
+                    "ConferenceTitle",
+                    "ConferenceDate",
+                    "ConferenceLocation",
+                    "ConferenceSponsors",
+                    "AuthorKeywords",
+                    "SubjectCategory",
+                    "TimesCited",
+                    "Abstract",
+                    "DOI",
+                    "Origin",
+                    "Label")
+
+test_that("get_reference_list works", {
+    rl <- get_reference_list(df)
+    # Number of rows
+    expect_equal(nrow(rl), number_of_rows)
+    # Full reference string
+    expect_equal(rl$FullReference[1], reference_1)
+    expect_equal(rl$FullReference[43], reference_43)
+    expect_equal(rl$FullReference[44], reference_44)
+    expect_equal(rl$FullReference[148], reference_148)
+    # Year
+    expect_equal(rl$ReferenceYear[1], year_1)
+    expect_equal(rl$ReferenceYear[43], year_43)
+    expect_equal(rl$ReferenceYear[44], year_44)
+    expect_equal(rl$ReferenceYear[148], year_148)
+    # DOI
+    expect_equal(rl$Reference[1], DOI_1)
+    expect_equal(rl$Reference[43], DOI_43)
+    expect_equal(rl$Reference[44], DOI_44)
+    expect_equal(rl$Reference[148], DOI_148)
+    # Missing values
+    expect_equal(sum(is.na(rl$Reference)), 0)
+})
+
+test_that("get_reference_nodes works", {
+    nodes <- get_reference_nodes(rl)
+    # Number of rows
+    expect_equal(nrow(nodes), number_of_rows - sum(duplicated(rl$Reference)))
+    # ID
+    expect_equal(nodes$Id[1], DOI_1)
+    expect_equal(nodes$Id[43], DOI_43)
+    expect_equal(nodes$Id[44], DOI_44)
+    expect_equal(nodes$Id[147], DOI_148)
+    # Year
[…]

[…]
