### 2017-07-12 22ee27e
Add tests to test_create_citation_network.R.
