### 2017-07-12 ba0c059
Add test_create_author_network.R.
