### 2017-07-14 fc9b3e8
Add author address extraction to create_author_network.R. All tests pass.
