### 2017-07-14 23a81b5
Add vignette.
