+++ DESCRIPTION
@@ -19,5 +19,8 @@ Imports:
     plyr,
     splitstackshape
 Suggests:
-    testthat
+    testthat,
+    knitr,
+    rmarkdown
 RoxygenNote: 6.0.1
+VignetteBuilder: knitr
