### 2017-07-16 12d0238
Vignette Hello World! Bug fix get_author_address.
