### 2017-07-18 785bf20
Add get_citation_network.
