### 2017-07-18 09fa701
Literature review vignette works except for Most referenced publications, topic model, and relative publication volume.
