+++ tests/testthat/test_create_author_network.R
@@ -2,11 +2,11 @@ library(nails)
 context("Creating an author network")
 
 # Read and clean test data
-df <- read_wos_txt(file.path("test_data/savedrecs-mactab.txt"), fix_names = TRUE)
+df <- read_wos_data(file.path("test_data"))
 df <- clean_wos_data(df)
 
 # Pick 3 rows for further processing
-df <- df[c(3, 36, 100), ]
+df <- df[c(3, 36, 100, 295), ]
 
 # Reference data
 author_id <- c("SACHDEVA, SONYA",
@@ -20,10 +20,15 @@ author_id <- c("SACHDEVA, SONYA",
                "RAAD, ELIE",
                "CHBEIR, RICHARD",
                "DIPANDA, ALBERT",
-               "RAAD, ELIANA J.")
+               "RAAD, ELIANA J.",
+               "JANAKI, NAFISEH",
+               "WHITNEY, JON",
+               "JANOWCZYK, ANDREW",
+               "MADABHUSHI, ANANT",
+               "AVRIL, STEFANIE")
 author_label <- author_id
-author_freq <- rep(1, 12)
-author_timescited <- rep(0, 12)
+author_freq <- rep(1, 17)
+author_timescited <- rep(0, 17)
 author_address <- c("US Forest Serv, 1033 Univ Pl,Ste 360, Evanston, IL 60201 USA",
                     "US Forest Serv, 1033 Univ Pl,Ste 360, Evanston, IL 60201 USA",
                     "Clark Univ, Grad Sch Geog, Worcester, MA 01610 USA",
@@ -35,16 +40,23 @@ author_address <- c("US Forest Serv, 1033 Univ Pl,Ste 360, Evanston, IL 60201 US
                     "Mem Univ Newfoundland, Fac Business, St John, NF A1C 5S7, Canada",
                     "Univ Pau, LIUPPA Lab, Pau, France",
                     "Univ Bourgogne, LE2I Lab, F-21004 Dijon, France",
-                    "Univ Bourgogne, LE2I Lab, F-21004 Dijon, France")
+                    "Univ Bourgogne, LE2I Lab, F-21004 Dijon, France",
+                    "Univ Hosp Cleveland Med Ctr, Cleveland, OH USA; Case Western Reserve Univ, Cleveland, OH 44106 USA",
+                    "Univ Hosp Cleveland Med Ctr, Cleveland, OH USA; Case Western Reserve Univ, Cleveland, OH 44106 USA",
+                    "Univ Hosp Cleveland Med Ctr, Cleveland, OH USA; Case Western Reserve Univ, Cleveland, OH 44106 USA",
+                    "Univ Hosp Cleveland Med Ctr, Cleveland, OH USA; Case Western Reserve Univ, Cleveland, OH 44106 USA",
+                    "Univ Hosp Cleveland Med Ctr, Cleveland, OH USA; Case Western Reserve Univ, Cleveland, OH 44106 USA")
 # author_reprintaddress not tested
 # Author email, improve processing!
 author_email <- c(rep("sonyasachdeva@fs.fed.us; smccaffrey@fs.fed.us; dexter.locke@gmail.com", 3),
                   rep("thomas.ludwig@uni-siegen.de", 5),
-                  rep("eliraad@gmail.com; richard.chbeir@univ-pau.fr; albert.dipanda@u-bourgogne.fr; eliana.raad@u-bourgogne.fr", 4))
+                  rep("eliraad@gmail.com; richard.chbeir@univ-pau.fr; albert.dipanda@u-bourgogne.fr; eliana.raad@u-bourgogne.fr", 4),
+                  rep("", 5))
 # Author location, improve processing!
 author_location <- c(rep("Evanston, IL 60201 USA;Worcester, MA 01610 USA", 3),
[…]
