+++ vignettes/literature_report.Rmd
@@ -15,7 +15,6 @@ vignette: >
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = FALSE, 
                       message=FALSE, 
-                      results='hide', 
                       warning=FALSE)
 ```
 
@@ -102,6 +101,9 @@ author_nodes <- transform(author_nodes,
                           AuthorFullName = reorder(AuthorFullName,
                                                    TotalTimesCited))
 
+# Sort by number of articles by author
+author_nodes <- author_nodes[with (author_nodes, order(-TotalTimesCited)), ]
+
 ggplot(head(author_nodes, 25), aes(AuthorFullName, TotalTimesCited)) +
     geom_bar(stat = "identity", fill = "blue") +
     coord_flip() +
@@ -130,7 +132,7 @@ publications <- transform(publications, Publication = reorder(Publication, Count
 
 
 # WHY???
-# literature <- merge(literature, citationSums,
+# literature <- merge(literature, citation_sums,
 #                    by = "PublicationName" )
 
 ggplot(head(publications, 25), aes(Publication, Count)) +
@@ -144,22 +146,22 @@ ggplot(head(publications, 25), aes(Publication, Count)) +
 
 ```{r}
 # Calculating total citations for each publication.
-citationSums <- aggregate(literature$TimesCited,
+citation_sums <- aggregate(literature$TimesCited,
     by = list(PublicationName = literature$PublicationName),
     FUN = sum, na.rm = T)
 
 # Fix column names
-names(citationSums) <- c("PublicationName", "PublicationTotalCitations")
+names(citation_sums) <- c("PublicationName", "PublicationTotalCitations")
 
 # Trim publication name to maximum of 50 characters for displaying in plot
-citationSums$PublicationName <- strtrim(citationSums$PublicationName, 50)
+citation_sums$PublicationName <- strtrim(citation_sums$PublicationName, 50)
 
 # Sort descending and reorder factor levels accordingly
-citationSums <- citationSums[with (citationSums, order(-PublicationTotalCitations)), ]
-citationSums <- transform(citationSums,
+citation_sums <- citation_sums[with (citation_sums, order(-PublicationTotalCitations)), ]
+citation_sums <- transform(citation_sums,
                           PublicationName = reorder(PublicationName,
                                                     PublicationTotalCitations))
-ggplot(head(citationSums, 25),
+ggplot(head(citation_sums, 25),
        aes(PublicationName, PublicationTotalCitations)) +
     geom_bar(stat = "identity", fill = "orange") +
     coord_flip() +
@@ -169,5 +171,133 @@ ggplot(head(citationSums, 25),
 
 ```
 
+## Important keywords
+Sorted by the number of articles where the keyword is mentioned and by the total number of citations for the keyword.
+```{r}
+# Calculating total citations for each keyword
+
+literature_by_keywords <- arrange_by(literature, "AuthorKeywords")
+
+
+
+# Sometimes AuthorKeywords column is empty.
+# Following if-else hack prevents crashing in those situations,
+# either by using KeywordsPlus column or skipping keyword analysis.
+if (nrow(literature_by_keywords) == 0) {
+  cat("No keywords.")
+} else {
+    keyword_citation_sum <- aggregate(literature_by_keywords$TimesCited,
+                                by = list(AuthorKeywords =
+                            literature_by_keywords$AuthorKeywords), FUN = sum,
+                            na.rm = T)
+    names(keyword_citation_sum) <- c("AuthorKeywords", "TotalTimesCited")
+    
+    keywords <- unlist(strsplit(literature$AuthorKeywords, ";"))
[…]
