### 2017-07-18 37e1ec8
Add testing create_citation_network.
