### 2017-07-18 7650d7d
Add get_author_network. Define functions to export (@export).
