### 2017-07-18 9f8125d
Add testing get_author_network, get_author_edges.
