### 2017-07-19 124b04d
Add most referenced publications, remove unnecessary library calls.
