### 2017-07-19 933eed6
Added Travis CI build settings
roxygen2 needed to build vignette
