### 2017-07-19 f2deb8f
Bug fix test_create_citation_network.R.
