### 2017-07-25 8ccb7a5
Added and edited the vignette
Added more structure, general information and topicmodels sample to the
vignette.
