+++ vignettes/literature_report.Rmd
@@ -1,6 +1,6 @@
 ---
-title: "Literature Review"
-author: "Juho Salminen"
+title: "Performing Literature Reviews with nails package"
+author: "Juho Salminen and Antti Knutas"
 date: "`r Sys.Date()`"
 output: 
     rmarkdown::html_vignette:
@@ -18,6 +18,18 @@ knitr::opts_chunk$set(echo = FALSE,
                       warning=FALSE)
 ```
 
+# Introduction
+NAILS performs statistical and Social Network Analysis (SNA) on citation data. SNA is a new way for researchers to map large datasets and get insights from new angles by analyzing connections between articles. As the amount of publications grows on any given field, automatic tools for this sort of analysis are becoming increasingly important prior to starting research on new fields. NAILS also provides useful data when performing Systematic Mapping Studies (SMS) in scientific literature. According to Kitchenham et al. performing a SMS can be especially suitable if few literature reviews have been done on the topic and there is a need to get a general overview of the field of interest.
+
+The nails package provides functionality for parsing Web of Science data for quantitative Systematic Mapping Study analysis, and a series of custom statistical and network analysis functions to give the user an overview of literature datasets. The features can be divided into two primary sections: Firstly, statistical analysis, which for example gives an overview of publication frequencies, most published authors and journals. Secondly, the more novel network analysis, which gives further insight into relationship between the interlinked citations and cooperation between authors. For example, the most basic features can use citation network analysis identify the most cited authors and publication forums. Lastly, it provides a few convenience functions to use the [topicmodels](https://cran.r-project.org/web/packages/topicmodels/index.html) and [stm](https://cran.r-project.org/web/packages/stm/index.html) packages to create Latent Dirichlet allocation -based [topic models](https://en.wikipedia.org/wiki/Topic_model).
+
+For further details see the following article: [Knutas, A., Hajikhani, A., Salminen, J., Ikonen, J., Porras, J., 2015. Cloud-Based Bibliometric Analysis Service for Systematic Mapping Studies. CompSysTech 2015](https://www.researchgate.net/publication/284644473_Cloud-Based_Bibliometric_Analysis_Service_for_Systematic_Mapping_Studies).
+
+# Example workflow and report
+In this section we present how to load Web of Science data using nails package functions and then to create an example report using [ggplot2](https://cran.r-project.org/web/packages/ggplot2/index.html)-based visualizations.
+
+## Loading data
+Below is an example of how data exported from Web of Science can be loaded and parsed using the nails package functions.
 
 ```{r}
 # Setup
@@ -36,14 +48,27 @@ literature <- read_wos_data("../tests/testthat/test_data")
 literature <- clean_wos_data(literature)
 ```
 
+## Generating visualizations with knittr
+Below we present how to generate example report using nails calls and then using ggplot2 and knittr to generate visual reports.
 
 This report provides an analysis on the records downloaded from [Web of Science](http://webofscience.com). The analysis identifies the important authors, journals, and keywords in the dataset based on the number of occurences and citation counts. A citation network of the provided records is created and used to identify the important papers according to their in-degree, total citation count and PageRank scores. The analysis finds also often-cited references that were not included in the original dataset downloaded from the Web of Science.
 
[…]
