### 2017-07-30 468d6cb
Fix bugs in tests.
