### 2017-07-30 2ab544c
Fix bug in vignette.
