### 2017-08-06 609432e
Added one reference to paper and silenced stm diagnostic output
