### 2017-08-06 #233 @aknutas
Dear editor @karthik ,
   @jsalminen and I have created a package version of the software at [https://github.com/aknutas/nails-package](https://github.com/aknutas/nails-package) . It has proper package structure and testthat-based tests that are automatically run by Travis CI. We have also written documentation as man files and in a vignette that explains how to create a report with the package.

Best regards,
   Antti Knutas and Juho Salminen
