### 2017-08-06 6486e62
Update NAMESPACE and documentation.
