+++ man/get_reference_list.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clean_data.R
+% Please edit documentation in R/create_citation_network.R
 \name{get_reference_list}
 \alias{get_reference_list}
 \title{Helper function for extracting reference lists}
