+++ man/get_all_fieldtags.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/update_fieldtags.R
+\name{get_all_fieldtags}
+\alias{get_all_fieldtags}
+\title{Download and merge all fieldtags from Web of Science}
+\usage{
+get_all_fieldtags(main_url, extra_url, extra_url_2)
+}
+\arguments{
+\item{main_url}{A URL for primary field tags}
+
+\item{extra_url}{A URL for supplementary field tags. Used for field tags missing in main_url.}
+
+\item{extra_url_2}{A URL for supplementary field tags. Used for field tags missing in extra_url.}
+}
+\value{
+A data frame containing all field tags
+}
+\description{
+Download and merge all fieldtags from Web of Science
+}
