+++ man/get_fieldtags.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/update_fieldtags.R
+\name{get_fieldtags}
+\alias{get_fieldtags}
+\title{Helper function for downloading fieldtags from Web of Science}
+\usage{
+get_fieldtags(url)
+}
+\arguments{
+\item{url}{A URL as a string}
+}
+\value{
+A data frame containing fieldtags downloaded from given URL
+}
+\description{
+Helper function for downloading fieldtags from Web of Science
+}
