+++ man/get_reference_nodes.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clean_data.R
+% Please edit documentation in R/create_citation_network.R
 \name{get_reference_nodes}
 \alias{get_reference_nodes}
 \title{Helper function for extracting citation network nodes from reference list}
