+++ man/get_citation_nodes.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clean_data.R
+% Please edit documentation in R/create_citation_network.R
 \name{get_citation_nodes}
 \alias{get_citation_nodes}
 \title{Extract nodes for a citation network from literature data}
