+++ man/get_citation_network.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/create_citation_network.R
+\name{get_citation_network}
+\alias{get_citation_network}
+\title{Extract citation network nodes and edges from literature data}
+\usage{
+get_citation_network(df, as_igraph = FALSE)
+}
+\arguments{
+\item{df}{A data frame containing literature data}
+
+\item{as_igraph}{Logical: should results be returned as an igraph?}
+}
+\value{
+A list containing nodes and edges in data frames, or an igraph.
+}
+\description{
+Extract citation network nodes and edges from literature data
+}
