+++ man/get_author_nodes.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/create_author_network.R
+\name{get_author_nodes}
+\alias{get_author_nodes}
+\title{Extract nodes for an author network from literature data}
+\usage{
+get_author_nodes(df)
+}
+\arguments{
+\item{df}{a data frame containing literature data}
+}
+\value{
+A data frame containing nodes for author network
+}
+\description{
+Extract nodes for an author network from literature data
+}
