### 2017-08-11 #233 @karthik
Hi @aknutas,
I will pass this back to @janschulz for additional review. In my quick review I did note some issues.

1. Please __do not__ include a packrat environment inside your package. Declare your dependencies. Packrat is meant for project level use and is not designed to be shipped with packages. Please address this issue asap.
2. Your README does not include any basic intructions on installation. Please add that in, and also link to a vignette.
3. Nice work on adding the `codemeta.json`
