+++ DESCRIPTION
@@ -17,7 +17,8 @@ Imports:
     dplyr,
     stringi,
     plyr,
-    splitstackshape
+    splitstackshape,
+    stringr
 Suggests:
     testthat,
     knitr,
