### 2017-12-17 daa6220
Literature report options set
