### 2017-12-17 b1fb0f8
Added more package requrements
