### 2017-12-17 e97e20c
Simplifying dependencies for Travis
