+++ .travis.yml
@@ -8,10 +8,8 @@ warnings_are_errors: false
 addons:
   apt:
     packages:
-    - gsl-bin
-    - libgsl0ldbl
-    - libgsl0-dev
-    - libgsl2
+  - gsl-bin
+  - libgsl0-dev
 
 r_packages:
   - covr
