### 2017-12-17 a5ce88e
Fixed Travis conf typo
