+++ .travis.yml
@@ -10,6 +10,7 @@ addons:
     packages:
     - gsl-bin
     - libgsl0-dev
+    - libgsl0ldbl
 
 r_packages:
   - covr
