### 2017-12-17 e0b726a
More Travis dependency testing (#3)
