+++ .travis.yml
@@ -8,10 +8,8 @@ warnings_are_errors: false
 addons:
   apt:
     packages:
-    - gsl-bin
     - libgsl0-dev
     - libgsl0ldbl
-    - libgsl2
 
 r_packages:
   - covr
