### 2017-12-17 aed8114
Updated README
New supported R version (tested on 3.4.2) and removed mention of
packrat.
