### 2017-12-17 14b72ab
Updated Travis configuration and added version requirements to DESCRIPTION file (#4)
* Literature report options set

* Added more package requrements

* Simplifying dependencies for Travis

* Fixed Travis conf typo

* More work with Travis

* More Travis debugging

* More Travis dependency testing (#3)
