+++ DESCRIPTION
@@ -9,19 +9,20 @@ License: GPL-3 | file LICENSE
 Encoding: UTF-8
 LazyData: true
 Imports:
-    stm,
-    tm,
-    LDAvis,
-    igraph,
-    topicmodels,
-    dplyr,
-    stringi,
-    plyr,
-    splitstackshape,
-    stringr
+    stm (>= 1.2.2),
+    tm (>= 0.7-1),
+    LDAvis (>= 0.3.2),
+    igraph (>= 1.1.2),
+    topicmodels (>= 0.2-6),
+    dplyr (>= 0.7.2),
+    stringi (>= 1.1.5),
+    plyr (>= 1.8.4),
+    splitstackshape (>= 1.4.2),
+    stringr (>= 1.2.0)
 Suggests:
-    testthat,
-    knitr,
-    rmarkdown
+    testthat (>= 1.0.2),
+    knitr (>= 1.17),
+    rmarkdown (>= 1.6),
+    roxygen2 (>= 6.0.1)
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
