### 2017-12-17 c35f8d4
Update vignette setup and formatting
Also added one more package to suggests
