+++ vignettes/literature_report.R
@@ -1,4 +1,0 @@
-## ----setup, include=FALSE------------------------------------------------
-knitr::opts_chunk$set(echo = FALSE,
-                      message = FALSE,
-                      warning = FALSE)
