### 2017-12-17 324435f
Updated project page and vignette docs
