+++ vignettes/literature_report_genpdf.Rmd
@@ -1,0 +1,369 @@
+---
+title: "Performing Literature Reviews with nails package"
+author: "Juho Salminen and Antti Knutas"
+date: "`r Sys.Date()`"
+output:
+    pdf_document:
+        fig_width: 7
+        fig_height: 5
+
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE, 
+                      message = FALSE, 
+                      warning = FALSE,
+		      tidy = TRUE,
+                      tidy.opts = list(width.cutoff=40))
+```
+
+# Introduction
+NAILS performs statistical and Social Network Analysis (SNA) on citation data. SNA is a new way for researchers to map large datasets and get insights from new angles by analyzing connections between articles. As the amount of publications grows on any given field, automatic tools for this sort of analysis are becoming increasingly important prior to starting research on new fields. NAILS also provides useful data when performing Systematic Mapping Studies (SMS) in scientific literature. According to Kitchenham et al. performing a SMS can be especially suitable if few literature reviews have been done on the topic and there is a need to get a general overview of the field of interest.
+
+The nails package provides functionality for parsing Web of Science data for quantitative Systematic Mapping Study analysis, and a series of custom statistical and network analysis functions to give the user an overview of literature datasets. The features can be divided into two primary sections: Firstly, statistical analysis, which for example gives an overview of publication frequencies, most published authors and journals. Secondly, the more novel network analysis, which gives further insight into relationship between the interlinked citations and cooperation between authors. For example, the most basic features can use citation network analysis identify the most cited authors and publication forums. Lastly, it provides a few convenience functions to use the [topicmodels](https://cran.r-project.org/web/packages/topicmodels/index.html) and [stm](https://cran.r-project.org/web/packages/stm/index.html) packages to create Latent Dirichlet allocation -based [topic models](https://en.wikipedia.org/wiki/Topic_model).
+
+For further details see the following article: [Knutas, A., Hajikhani, A., Salminen, J., Ikonen, J., Porras, J., 2015. Cloud-Based Bibliometric Analysis Service for Systematic Mapping Studies. CompSysTech 2015](https://www.researchgate.net/publication/284644473_Cloud-Based_Bibliometric_Analysis_Service_for_Systematic_Mapping_Studies).
+
+# Example workflow and report
+In this section we present how to load Web of Science data using nails package functions and then to create an example report using [ggplot2](https://cran.r-project.org/web/packages/ggplot2/index.html)-based visualizations.
+
+## Loading data
+Below is an example of how data exported from Web of Science can be loaded and parsed using the nails package functions.
+
+```{r}
+# Setup
+
+# Load packages
+devtools::load_all()
+require(ggplot2)
+
+# Set ggplot theme
+theme_set(theme_minimal(12))
+
+# Load data
+literature <- read_wos_data("../tests/testthat/test_data")
+
+# Clean data
+literature <- clean_wos_data(literature)
+```
+
+## Generating visualizations with knittr
+Below we present how to generate example report using nails calls and then using ggplot2 and knittr to generate visual reports.
+
[…]
