### 2017-12-17 0161b87
Corrected vignette path
