### 2017-12-17 d1e6cf7
Updated README
Details on how to get started
