### 2017-12-17 bc7c1d2
A small correction to how to use
