### 2017-12-18 e41952d
Updating the main branch from develop improvements (#5)
* Literature report options set

* Added more package requrements

* Simplifying dependencies for Travis

* Fixed Travis conf typo

* More work with Travis

* More Travis debugging

* More Travis dependency testing (#3)

* Updated README

New supported R version (tested on 3.4.2) and removed mention of
packrat.

* Update vignette setup and formatting

Also added one more package to suggests

* Updated README

Details on how to get started
