+++ vignettes/literature_report.Rmd
@@ -12,10 +12,11 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = FALSE, 
-                      message=FALSE, 
-                      warning=FALSE)
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(echo = TRUE, 
+                      message = FALSE, 
+                      warning = FALSE,
+                      tidy = TRUE)
 ```
 
 # Introduction
