+++ DESCRIPTION
@@ -23,6 +23,7 @@ Suggests:
     testthat (>= 1.0.2),
     knitr (>= 1.17),
     rmarkdown (>= 1.6),
-    roxygen2 (>= 6.0.1)
+    roxygen2 (>= 6.0.1),
+    formatR (>= 1.5)
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
