### 2017-12-18 #233 @aknutas
Dear editor @karthik and reviewer @janschulz ,
   thank you for your feedback and the patience for first time package builders. We have now improved the package accordingly.

1. Packrat has now been moved to development branch and the main branch is cleaned and "ready for public". The [DESCRIPTION file](https://github.com/aknutas/nails-package/blob/master/DESCRIPTION) has now the required version numbering added to the list of dependencies. We have verified that the listed dependencies are sufficient (for example Travis CI pulls required R packages using that).
2. README has now [use instructions](https://github.com/aknutas/nails-package#how-to-use) and links to the vignette. The [vignette](http://anttiknutas.net/nails-package/vignettes/literature_report.html) has been prepared with through usage examples that include sample code and sample output.

Additionally we have [automatic test cases](https://github.com/aknutas/nails-package/tree/master/tests) for the functions, and they have been set up to be checked by [Travis CI automatically](https://travis-ci.org/aknutas/nails-package).

We hope the improvements are sufficient. This has been a great learning experience and we look forward to instructions for finishing the review process.

Best regards,
   Antti Knutas and Juho Salminen
