### 2018-02-27 #233 @jankatins
I only looked at the issues left open in https://github.com/openjournals/joss-reviews/issues/233#issuecomment-292734519

# repository URL

This is now based on https://github.com/aknutas/nails-package, so the mentoned repository URL in the first coment is NOT correct anymore. :question:

# version

The package has a 1.0.2 release: https://github.com/aknutas/nails-package/releases/tag/1.0.2 so the version is not the same as the one mentioned in the first comment anymore :-( :question:

# automated tests

testthat tests are there and travis is enabled :white_check_mark:

# citations

Looks great now! :white_check_mark:

I ran the vignette and this looks nice.

So apart from the change of the repo and version, this looks good. I'm not sure how to solve that problem. @karthik, can you help out with that?
