### 2018-02-27 aknutas/nails-package#7 @jankatins
Add function to generate a rmarkdown report from a dataset and/or highlevel wrappers 
The old way was downloading a dataste and then rerunnign a rmarkdown file. Now you have to do this all by yourself or copy the vignette code. It would be nice if the old report could be taken as a template for a `generate_report(datafile, reportfilename)` functions which would

1. generate a rmarkdown file
2. knit the file and show it in the browser
3. spit out some help text so that it's easy to change and rerun the report

An alternative (or additionally) would be to add some highlevel functions to spit out graphs/tables instead of letting the user do the data munching themselves (the vignette has quite a lot of code instead of one command per graph/table).

(this came out of my review in https://github.com/openjournals/joss-reviews/issues/233#issuecomment-369065201)
