### 2018-02-27 aknutas/nails-package#8 @jankatins
Add function to generate a "historic view" on the filed
Basically I want this:
* citation graph of the most important papers
* Arrange the y axis according to paper year
* links  = citations
* label thickness of each node = (citation count / age)?
* Colour = clustered or topicmodels?

```
        paper 1950
    /                  \
paper 1970            paper 1980
    \                  /
        paper 1990
```

This can be generated by igrpah by limiting the y axis positions to predefined values (at least in python).

This gives a nice overview what paper you should read...

This is the same issue as https://github.com/aknutas/nails/issues/14 (which has example code to generate such graphs, but in python)
