### 2018-02-28 aknutas/nails-package#8 @aknutas
Thank you for the feedback, @jankatins ! Now that the refactoring is completed, we will start working for this and other view features for the next iteration of the library.
