### 2018-02-28 #233 @aknutas
Thank you for the feedback, @jankatins ! Now that the refactoring is completed, we look forward to implementing additional usability and analysis features in the near future. Your comments and the examples provided will help us retarget efforts for the 1.1.0 release.

We had to change the repository, because the old structure of the R package was very nonstandard and not a package at all. Additionally, we received feedback during the review process that had us updating releases, which is why the release version increased. Could the editor @karthik offer advice regarding this?
