### 2018-03-27 #233 @aknutas
Thank you for the further information. The new repository link is available at https://github.com/aknutas/nails-package and the release is https://github.com/aknutas/nails-package/releases/tag/1.0.2 . @jankatins can you edit the original issue as I don't have edit access?
