### 2018-04-03 #233 @karthik
@aknutas

I found a few more issues that I think you should fix before we accept
```

  ✖ add a "URL" field to DESCRIPTION. It helps users find
    information about your package online. If your package does not
    have a homepage, add an URL to GitHub, or the CRAN package package
    page.```

Example: https://github.com/karthik/wesanderson/blob/master/DESCRIPTION#L20

```✖ add a "BugReports" field to DESCRIPTION, and point it to
    a bug tracker. Many online code hosting services provide bug
    trackers for free, https://github.com, https://gitlab.com, etc.```

Example: https://github.com/karthik/wesanderson/blob/master/DESCRIPTION#L21

 ``` ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/clean_data.R:108:10

    R/create_author_network.R:10:1
    R/create_author_network.R:70:1
    R/create_author_network.R:75:1
    R/create_author_network.R:76:1
    R/create_author_network.R:150:1
    ... and 49 more lines

    vapply() instead.

    R/clean_data.R:72:20
    R/clean_data.R:157:19
    R/clean_data.R:159:17
    R/clean_data.R:161:20
    R/create_author_network.R:120:20
    ... and 2 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/clean_data.R:77:14
    R/create_author_network.R:80:19
    R/topic_model.R:89:15

  ✖ fix this R CMD check WARNING: Found the following file
    with a non-portable file name:
    tests/testthat/test_data/datafolder/crowdsourcing_old (1).txt These
    are not fully portable file names. See section ‘Package structure’
    in the ‘Writing R Extensions’ manual.

Don't use file names with spaces in the data folder.

  ✖ fix this R CMD check WARNING: Found the following file
    with non-ASCII characters: read_data.R Portable packages must use
    only ASCII characters in their R code, except perhaps in comments.
    Use \uxxxx escapes for other characters.

Remove UTF-8 characters

  ✖ fix this R CMD check WARNING: '::' or ':::' imports not
    declared from: ‘devtools’ ‘RCurl’ ‘XML’ Namespaces in Imports field
    not imported from: ‘dplyr’ ‘stringi’ All declared Imports should be
    used.

Remove any undeclared imports

  ✖ fix this R CMD check NOTE: clean_wos_data: no visible
    binding for global variable ‘fixed_fieldtags’ create_edges: no
    visible global function definition for ‘combn’ get_author_edges: no
    visible binding for global variable ‘Authors’ get_author_edges: no
    visible binding for global variable ‘AuthorFullName’
    get_author_nodes: no visible global function definition for
    ‘aggregate’ get_author_nodes: no visible binding for global
    variable ‘Id’ get_author_nodes: no visible binding for global
    variable ‘Label’ get_author_nodes: no visible binding for global
    variable ‘Freq’ get_author_nodes: no visible binding for global
    variable ‘TotalTimesCited’ get_author_nodes: no visible binding for
    global variable ‘AuthorAddress’ get_author_nodes: no visible
    binding for global variable ‘EmailAddress’ get_author_nodes: no
    visible binding for global variable ‘Location’
    get_literature_nodes: no visible binding for global variable ‘DOI’
    get_literature_nodes: no visible binding for global variable
    ‘YearPublished’ get_literature_nodes: no visible binding for global
    variable ‘ReferenceString’ get_literature_nodes: no visible binding
    for global variable ‘id’ get_literature_nodes: no visible binding
    for global variable ‘PublicationType’ get_literature_nodes: no
    visible binding for global variable ‘AuthorFullName’
    get_literature_nodes: no visible binding for global variable
    ‘DocumentTitle’ get_literature_nodes: no visible binding for global
    variable ‘PublicationName’ get_literature_nodes: no visible binding
    for global variable ‘BookSeriesTitle’ get_literature_nodes: no
    visible binding for global variable ‘Language’
    get_literature_nodes: no visible binding for global variable
    ‘DocumentType’ get_literature_nodes: no visible binding for global
    variable ‘ConferenceTitle’ get_literature_nodes: no visible binding
    for global variable ‘ConferenceDate’ get_literature_nodes: no
    visible binding for global variable ‘ConferenceLocation’
    get_literature_nodes: no visible binding for global variable
    ‘ConferenceSponsors’ get_literature_nodes: no visible binding for
    global variable ‘AuthorKeywords’ get_literature_nodes: no visible
    binding for global variable ‘SubjectCategory’ get_literature_nodes:
    no visible binding for global variable ‘TimesCited’
    get_literature_nodes: no visible binding for global variable
    ‘Abstract’ get_location : <anonymous>: no visible global function
    definition for ‘tail’ read_wos_folder: no visible global function
    definition for ‘read.delim2’ read_wos_folder: no visible binding
    for global variable ‘read.delim2’ read_wos_txt: no visible global
    function definition for ‘read.delim2’ topicmodels_json_ldavis: no
    visible global function definition for ‘%>%’
    topicmodels_json_ldavis: no visible global function definition for
    ‘stri_count’ Undefined global functions or variables: %>% Abstract
    aggregate AuthorAddress AuthorFullName AuthorKeywords Authors
    BookSeriesTitle combn ConferenceDate ConferenceLocation
    ConferenceSponsors ConferenceTitle DocumentTitle DocumentType DOI
    EmailAddress fixed_fieldtags Freq id Id Label Language Location
    PublicationName PublicationType read.delim2 ReferenceString
    stri_count SubjectCategory tail TimesCited TotalTimesCited
    YearPublished Consider adding importFrom("stats", "aggregate")
    importFrom("utils", "combn", "read.delim2", "tail") to your
    NAMESPACE file.
  ✖ fix this R CMD check WARNING: Undocumented code objects:
    ‘fixed_fieldtags’ ‘original_fieldtags’ Undocumented data sets:
    ‘fixed_fieldtags’ ‘original_fieldtags’ All user-level objects in a
    package should have documentation entries. See chapter ‘Writing R
    documentation files’ in the ‘Writing R Extensions’ manual.
  ✖ fix this R CMD check WARNING: Warning: package needs
    dependence on R (>= 2.10)
  ✖ fix this R CMD check WARNING: Warning: package needs
    dependence on R (>= 2.10) fix this R CMD check WARNING: Warning:
    package needs dependence on R (>= 2.10)
  ✖ fix this R CMD check NOTE: '::' or ':::' import not
    declared from: ‘devtools’ 'library' or 'require' call not declared
    from: ‘ggplot2’
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no
    redirection of stdout/stderr. Hmm ... looks like a package You may
    want to clean up by 'rm -rf /tmp/RtmpkunlEX/Rd2pdf759d1c0d9b97'
  ✖ avoid 'T' and 'F', as they are just variables which are
    set to the logicals 'TRUE' and 'FALSE' by default, but are not
    reserved words and hence can be overwritten by the user.  Hence,
    one should always use 'TRUE' and 'FALSE' for the logicals.

    R/create_author_network.R:NA:NA```
