### 2018-06-15 #233 @arfon
@jsalminen @aknutas - please respond to @karthik's feedback.
