### 2018-06-16 #233 @aknutas
Dear @karthik and @arfon,
   apologies for the delay. After conferring, we have one query about the review process. For acceptance, is it enough that we address the issues mentioned [https://github.com/openjournals/joss-reviews/issues/233#issuecomment-378091508](https://github.com/openjournals/joss-reviews/issues/233#issuecomment-378091508); or is it also required to implement the extra features mentioned in [https://github.com/aknutas/nails-package/issues/7](https://github.com/aknutas/nails-package/issues/7) and [https://github.com/aknutas/nails-package/issues/8](https://github.com/aknutas/nails-package/issues/8) ?

Best regards,
   Antti Knutas and Juho Salminen
