### 2018-08-10 #233 @karthik
@aknutas Just addressing these outstanding issues would be sufficient. 🙏
