### 2018-09-09 869b952
Web of Science data format and field tag change fix.
