+++ tests/testthat/test_create_citation_network.R
@@ -41,7 +41,6 @@ nodes_colnames <- c("Id",
                     "ConferenceLocation",
                     "ConferenceSponsors",
                     "AuthorKeywords",
-                    "SubjectCategory",
                     "TimesCited",
                     "Abstract",
                     "DOI",
