+++ R/topic_model.R
@@ -2,7 +2,8 @@
 #' Use for select_optimal_k or others that take the "data$textlist" object.
 #'
 #' @param literature The cleaned and processed literature dataframe
-#' @return A list with vector of words (textlist) and data frame with only abstracts, topics and rowids (doctablewt)
+#' @return A list with vector of words (textlist) and data frame with only
+#' abstracts, topics and rowids (doctablewt)
 #'
 #' @export
 
@@ -14,7 +15,9 @@ preprocess_literature_for_topicmodeling <- function(literature) {
     literature$topicmodelrowids <-as.numeric(rownames(literature))
 
     # Subselect dataframe for processing and drop empty rows
-    doctablewt <- literature[,c("DocumentTitle", "Abstract", "topicmodelrowids")]
+    doctablewt <- literature[,c("DocumentTitle",
+                                "Abstract",
+                                "topicmodelrowids")]
     doctablewt <- doctablewt[!is.na(doctablewt$Abstract), ]
     doctablewt <- doctablewt[!is.na(doctablewt$DocumentTitle), ]
 
@@ -31,30 +34,37 @@ preprocess_literature_for_topicmodeling <- function(literature) {
 build_stm_corpus <- function(textlist) {
     # Prepare documents into corpus
     # Also processes the following by default:
-    # Lowercases, removes SMART stopwords, removes numbers, removes punctuation, wordlength min is 3
+    # Lowercases, removes SMART stopwords, removes numbers, removes punctuation,
+    # wordlength min is 3
     processed <- stm::textProcessor(textlist, stem = TRUE, verbose = FALSE)
-    out <- stm::prepDocuments(processed$documents, processed$vocab, verbose = FALSE)
+    out <- stm::prepDocuments(processed$documents, processed$vocab,
+                              verbose = FALSE)
     return(out)
 }
 
 #' Estimates number of topics (K) using the stm library
 #'
 #' @param data Data from preprocess_literature_for_topicmodeling()
-#' @return A list with topic estimation from stm lib (topickest), semantic coherence values (semcohvalues) and estimate of best K of topics (bestk)
+#' @return A list with topic estimation from stm lib (topickest),
+#' semantic coherence values (semcohvalues)
+#' and estimate of best K of topics (bestk)
 #'
 #' @export
 
 select_optimal_k <- function(data) {
     out <- build_stm_corpus(data$textlist)
 
-    # Estimate number of topics; semantic coherence often good (default method spectral; best compromise and deterministic)
+    # Estimate number of topics; semantic coherence often good
+    # (default method spectral; best compromise and deterministic)
     # Seed set for consitent results
     set.seed(5707363)
-    topickest <- stm::searchK(out$documents, out$vocab, K = c(4:12), seed = 5707363, verbose = FALSE)
+    topickest <- stm::searchK(out$documents, out$vocab, K = c(4:12),
+                              seed = 5707363, verbose = FALSE)
     semcohsK <- data.frame(topickest$results$K, topickest$results$semcoh)
     colnames(semcohsK)<- c("K","semcohs")
 
-    # Getting the K with highest semantic coherence, setting it as true in semcohsK DF
+    # Getting the K with highest semantic coherence,
+    # setting it as true in semcohsK DF
     bestpick <- semcohsK[which.max(semcohsK$semcohs),]
     bestK <- as.integer(bestpick$K)
 
@@ -90,7 +100,10 @@ topicmodels_json_ldavis <- function(fitted, corpus, doc_term){
         temp <- paste(corpus[[i]]$content, collapse = ' ')
         doc_length <- c(doc_length, stri_count(temp, regex = '\\S+'))
     }
[…]
