### 2018-09-16 a95d8dd
Use vapply() instead of sapply().
