### 2018-09-16 55b6cbc
Replace length() and nrow() with seq_along()
